+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v3.10.0-dev
 
 - Enhancements
+    - Upstream improvements for "360 Secure" browser detection have been integrated
     - Upstream improvements for "PICO OS" detection have been integrated
     - Upstream improvements for "SPARC64" platform detection have been integrated